% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kansuji2arabic_MeCab.R
\name{kansuji2arabic_str_MeCab}
\alias{kansuji2arabic_str_MeCab}
\title{Convert kansuji to Arabic numerals with MeCab}
\usage{
kansuji2arabic_str_MeCab(
  str,
  consecutive = c("convert", "non"),
  widths = c("all", "halfwidth"),
  ...
)
}
\arguments{
\item{str}{Input vector.}

\item{consecutive}{If you select "convert", any sequence of 1 to 9 kansuji will
be replaced with Arabic numerals. If you select "non", any sequence of 1-9
kansuji will not be replaced by Arabic numerals.}

\item{widths}{If you select "all", both full-width and half-width Arabic numerals
are taken into account when calculating kansuji, but if you select "halfwidth",
only half-width Arabic numerals are taken into account when calculating kansuji.}

\item{...}{Other arguments to carry over.}
}
\value{
a character.
}
\description{
Converts a given kansuji to arabic mumerals with MeCab.
`kansuji2arabic_str()` converts kansuji in a string to numbers represented by kansuji while
retaining the non-kansuji characters with MeCab.
}
